+++ src/main/java/subway/application/StationService.java
@@ -4,8 +4,8 @@ import java.util.List;
 import java.util.stream.Collectors;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import subway.controller.dto.StationRequest;
-import subway.controller.dto.StationResponse;
+import subway.controller.dto.request.StationRequest;
+import subway.controller.dto.response.StationResponse;
 import subway.domain.Station;
 import subway.exception.BusinessException;
 import subway.persistence.StationRepository;
