+++ src/main/java/subway/controller/LineStationController.java
@@ -7,12 +7,12 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import subway.application.LineStationService;
-import subway.controller.dto.AddInitStationToLineRequest;
 import subway.controller.dto.AddStationLocation;
-import subway.controller.dto.AddStationToBetweenLineRequest;
-import subway.controller.dto.AddStationToEndLineRequest;
-import subway.controller.dto.AddStationToLineRequest;
-import subway.controller.dto.RemoveStationOnLineRequest;
+import subway.controller.dto.request.AddInitStationToLineRequest;
+import subway.controller.dto.request.AddStationToBetweenLineRequest;
+import subway.controller.dto.request.AddStationToEndLineRequest;
+import subway.controller.dto.request.AddStationToLineRequest;
+import subway.controller.dto.request.RemoveStationOnLineRequest;
 
 @RestController
 @RequestMapping("/line/station")
