+++ src/main/java/subway/controller/PathController.java
@@ -5,8 +5,8 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 import subway.application.PathService;
-import subway.controller.dto.FindShortestPathRequest;
-import subway.controller.dto.FindShortestPathResponse;
+import subway.controller.dto.request.FindShortestPathRequest;
+import subway.controller.dto.response.FindShortestPathResponse;
 
 @RestController
 public class PathController {
@@ -18,7 +18,8 @@ public class PathController {
     }
 
     @GetMapping("/shortest-path")
-    public ResponseEntity<FindShortestPathResponse> findShortestPath(@RequestBody final FindShortestPathRequest request) {
+    public ResponseEntity<FindShortestPathResponse> findShortestPath(
+        @RequestBody final FindShortestPathRequest request) {
         final FindShortestPathResponse response = pathService.findShortestPath(request);
         return ResponseEntity.ok().body(response);
     }
