+++ src/main/java/subway/controller/dto/request/AddStationToLineRequest.java
@@ -1,18 +1,28 @@
-package subway.controller.dto;
+package subway.controller.dto.request;
 
-public class AddStationToBetweenLineRequest {
+import javax.validation.constraints.NotBlank;
+import subway.controller.dto.AddStationLocation;
 
+public class AddStationToLineRequest {
+
+    @NotBlank
+    private AddStationLocation addStationLocation;
+    @NotBlank
     private String lineName;
+    @NotBlank
     private String stationName;
     private String upStationName;
     private String downStationName;
+    @NotBlank
     private Long distance;
 
-    private AddStationToBetweenLineRequest() {
+    private AddStationToLineRequest() {
     }
 
-    public AddStationToBetweenLineRequest(final String lineName, final String stationName, final String upStationName,
-        final String downStationName, final Long distance) {
+    public AddStationToLineRequest(final AddStationLocation addStationLocation, final String lineName,
+        final String stationName,
+        final String upStationName, final String downStationName, final Long distance) {
+        this.addStationLocation = addStationLocation;
         this.lineName = lineName;
         this.stationName = stationName;
         this.upStationName = upStationName;
@@ -20,10 +30,8 @@ public class AddStationToBetweenLineRequest {
         this.distance = distance;
     }
 
-    public static AddStationToBetweenLineRequest from(final AddStationToLineRequest addStationToLineRequest) {
-        return new AddStationToBetweenLineRequest(addStationToLineRequest.getLineName(),
-            addStationToLineRequest.getStationName(), addStationToLineRequest.getUpStationName(),
-            addStationToLineRequest.getDownStationName(), addStationToLineRequest.getDistance());
+    public AddStationLocation getAddStationLocation() {
+        return addStationLocation;
     }
 
     public String getLineName() {
