+++ src/test/java/subway/application/LineStationServiceTest.java
@@ -14,10 +14,10 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
-import subway.controller.dto.AddInitStationToLineRequest;
-import subway.controller.dto.AddStationToBetweenLineRequest;
-import subway.controller.dto.AddStationToEndLineRequest;
-import subway.controller.dto.RemoveStationOnLineRequest;
+import subway.controller.dto.request.AddInitStationToLineRequest;
+import subway.controller.dto.request.AddStationToBetweenLineRequest;
+import subway.controller.dto.request.AddStationToEndLineRequest;
+import subway.controller.dto.request.RemoveStationOnLineRequest;
 import subway.domain.Line;
 import subway.domain.Section;
 import subway.domain.Sections;
