+++ src/test/java/subway/application/PathServiceTest.java
@@ -15,9 +15,9 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
-import subway.controller.dto.FindShortestPathRequest;
-import subway.controller.dto.FindShortestPathResponse;
-import subway.controller.dto.StationResponse;
+import subway.controller.dto.request.FindShortestPathRequest;
+import subway.controller.dto.response.FindShortestPathResponse;
+import subway.controller.dto.response.StationResponse;
 import subway.domain.CostPolicy;
 import subway.domain.Line;
 import subway.domain.Section;
