+++ src/test/java/subway/integration/LineStationIntegrationTest.java
@@ -12,13 +12,13 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
-import subway.controller.dto.AddInitStationToLineRequest;
 import subway.controller.dto.AddStationLocation;
-import subway.controller.dto.AddStationToLineRequest;
-import subway.controller.dto.LineRequest;
-import subway.controller.dto.LineResponse;
-import subway.controller.dto.RemoveStationOnLineRequest;
-import subway.controller.dto.StationResponse;
+import subway.controller.dto.request.AddInitStationToLineRequest;
+import subway.controller.dto.request.AddStationToLineRequest;
+import subway.controller.dto.request.LineRequest;
+import subway.controller.dto.request.RemoveStationOnLineRequest;
+import subway.controller.dto.response.LineResponse;
+import subway.controller.dto.response.StationResponse;
 
 public class LineStationIntegrationTest extends IntegrationTest {
